+++ GUI.py
@@ -1,4 +1,4 @@
-from qom.ui import run
+from qom.ui.gui import run
 
 if __name__ == '__main__':
     run()