+++ scripts/bi_00_v=mode_amp.py
@@ -10,27 +10,26 @@ from qom.ui.plotters import MPLPlotter
 # add path to local libraries
 sys.path.append(os.path.abspath(os.path.join('..', 'sync_bi_uni')))
 # import system
-from systems.Bi00 import Bi00
+from systems import Bi00
 
 # all parameters
 params = {
     'solver': {
         'show_progress': True,
         'cache': True,
-        'cache_dir': 'H:/Workspace/data/bi_00/0.0_1000.0_10001',
-        'method': 'ode',
+        'method': 'zvode',
         'measure_type': 'mode_amp',
-        'idx_mode': [1, 3],
-        'range_min': 0,
-        'range_max': 10001,
-        't_min': 0,
-        't_max': 1000,
+        'idx_e': [1, 3],
+        'range_min': 4900,
+        'range_max': 5001,
+        't_min': 0.0,
+        't_max': 1000.0,
         't_dim': 10001
     },
     'system': {
         'A_l': 52.0,
         'Delta_0': 1.0, 
-        'delta': 0.0, 
+        'delta': 0.01, 
         'g_0s': [0.005, 0.005],
         'gammas': [0.005, 0.005],
         'kappas': [0.15, 0.15],
@@ -40,18 +39,15 @@ params = {
     },
     'plotter': {
         'type': 'lines',
-        'palette': 'blr',
-        'bins': 2,
         'x_label': '$\\omega_{mL} t$',
-        'x_bound': 'both',
         'x_ticks': [490, 495, 500],
-        'v_bound': 'both',
         'v_ticks': [-400, 0, 400],
         'show_legend': True,
-        'legend_location': 'lower right', 
-        'y_legend': ['$\\langle b_{L} \\rangle_{\\mathrm{Re}}$', '$\\langle b_{R} \\rangle_{\\mathrm{Re}}$'],
+        'legend_location': 'upper right',
         'label_font_size': 44,
-        'tick_font_size': 36
+        'tick_font_size': 36,
+        'width': 7.5,
+        'height': 6.0
     }
 }
 
@@ -59,22 +55,16 @@ params = {
 init_log()
 
 # initialize system
-system = Bi00(params['system'])
+system = Bi00(params=params['system'])
 
-# elements
-M_ele, T = system.get_measure_dynamics(params['solver'], system.ode_func, system.get_ivc)
-M_ele = np.imag(M_ele)
+# get mode amplitudes
+M, T = system.get_measure_dynamics(solver_params=params['solver'])
+M_real = np.real(M)
 
 # plotter
-axes = {
+plotter = MPLPlotter(axes={
     'X': T,
-    'Y': {
-        'var': 'b',
-        'val': [0, 1]
-    }
-}
-plotter = MPLPlotter(axes, params['plotter'])
-_xs = [T, T]
-_vs = np.transpose(M_ele).tolist()
-plotter.update(xs=_xs, vs=_vs)
-plotter.show(True, 7.5, 6.0)
+    'Y': ['$\\langle b_{L} \\rangle_{\\mathrm{Re}}$', '$\\langle b_{R} \\rangle_{\\mathrm{Re}}$']
+}, params=params['plotter'])
+plotter.update(xs=T, vs=np.transpose(M_real))
+plotter.show(True)