+++ scripts/bi_00_x=q1_y=p1_v=Wigner.py
@@ -1,4 +1,5 @@
 # dependencies
+import itertools
 import numpy as np
 import os 
 import sys
@@ -10,49 +11,45 @@ from qom.ui.plotters import MPLPlotter
 # add path to local libraries
 sys.path.append(os.path.abspath(os.path.join('..', 'sync_bi_uni')))
 # import system
-from systems.Bi00 import Bi00
+from systems import Bi00
 
 # parameters
 params = {
     'solver': {
         'show_progress': True,
         'cache': True,
-        'cache_dir': 'H:/Workspace/data/bi_00/0.0_1000.0_10001',
-        'method': 'ode',
+        'method': 'zvode',
         'measure_type': 'corr_ele',
-        'idx_row': [2, 2, 2, 2, 3, 3, 3, 3, 6, 6, 6, 6, 7, 7, 7, 7],
-        'idx_col': [2, 3, 6, 7, 2, 3, 6, 7, 2, 3, 6, 7, 2, 3, 6, 7],
+        'idx_e': list(itertools.product([2, 3, 6, 7], [2, 3, 6, 7])),
         'range_min': 9371,
         'range_max': 10001,
-        't_min': 0,
-        't_max': 1000,
+        't_min': 0.0,
+        't_max': 1000.0,
         't_dim': 10001
     },
     'system': {
         'A_l': 52.0,
         'Delta_0': 1.0, 
-        'delta': 0.0116, 
+        'delta': 0.01,
         'g_0s': [0.005, 0.005],
         'gammas': [0.005, 0.005],
         'kappas': [0.15, 0.15],
-        'lamb': 0.05,
+        'lamb': 0.075,
         'n_ths': [0, 0],
         'omega_m': 1.0
     },
     'plotter': {
         'type': 'contourf',
-        'palette': 'RdBu_r',
-        'bins': 11,
-        'title': '$\\delta = 0.0116, \\lambda = 0.05$',
+        'title': '$\\delta = 0.005, \\lambda = 0.075$',
         'x_label': '$\\langle q_{R} \\rangle$',
-        'x_bound': 'both',
-        'x_ticks': [-20, 0, 20],
+        'x_ticks': [-5, 0, 5],
         'y_label': '$\\langle p_{R} \\rangle$',
-        'y_bound': 'both',
-        'y_ticks': [-20, 0, 20],
-        'cbar_ticks': [0, 0.0005],
-        'label_font_size': 33,
-        'tick_font_size': 27
+        'y_ticks': [-5, 0, 5],
+        'cbar_title': '$\\times 10^{-3}$',
+        'cbar_ticks': [0.0, 0.0015, 0.003],
+        'cbar_tick_labels': [0.0, 1.5, 3.0],
+        'width': 5.5,
+        'height': 5.0
     }
 }
 
@@ -60,20 +57,20 @@ params = {
 init_log()
 
 # initialize system
-system = Bi00(params['system'])
+system = Bi00(params=params['system'])
 
-# get dynamics
-M, _ = system.get_measure_dynamics(params['solver'], system.ode_func, system.get_ivc)
+# get correlation elements
+M, _ = system.get_measure_dynamics(solver_params=params['solver'])
 V_mech = [np.reshape(m, (4, 4)) for m in M[-1:]]
 
 # initialize variables
 q_0 = 0
 p_0 = 0
-q_1s = np.around(np.linspace(-20, 20, 101), 1)
-p_1s = np.around(np.linspace(-20, 20, 101), 1)
+q_1s = np.around(np.linspace(-5, 5, 101), 1)
+p_1s = np.around(np.linspace(-5, 5, 101), 1)
 
 for V in V_mech:
-    # wigner
+    # wigner function
     W = list()
     for q_1 in q_1s:
         temp = list()
@@ -84,10 +81,9 @@ for V in V_mech:
         W.append(temp)
 
     # plotter
-    axes = {
+    plotter = MPLPlotter(axes={
         'X': q_1s.tolist(),
         'Y': p_1s.tolist()
-    }
-    plotter = MPLPlotter(axes, params['plotter'])
+    }, params=params['plotter'])
     plotter.update(vs=W)
-    plotter.show(True, 6.0, 5.0)
+    plotter.show(True)