+++ scripts/uni_00_x=p0_y=p1_v=Wigner.py
@@ -1,4 +1,5 @@
 # dependencies
+import itertools
 import numpy as np
 import os 
 import sys
@@ -10,22 +11,20 @@ from qom.ui.plotters import MPLPlotter
 # add path to local libraries
 sys.path.append(os.path.abspath(os.path.join('..', 'sync_bi_uni')))
 # import system
-from systems.Uni00 import Uni00
+from systems import Uni00
 
 # parameters
 params = {
     'solver': {
         'show_progress': True,
         'cache': True,
-        'cache_dir': 'H:/Workspace/data/uni_00/0.0_10000.0_100001',
-        'method': 'ode',
+        'method': 'zvode',
         'measure_type': 'corr_ele',
-        'idx_row': [2, 2, 2, 2, 3, 3, 3, 3, 6, 6, 6, 6, 7, 7, 7, 7],
-        'idx_col': [2, 3, 6, 7, 2, 3, 6, 7, 2, 3, 6, 7, 2, 3, 6, 7],
+        'idx_e': list(itertools.product([2, 3, 6, 7], [2, 3, 6, 7])),
         'range_min': 99371,
         'range_max': 100001,
-        't_min': 0,
-        't_max': 10000,
+        't_min': 0.0,
+        't_max': 10000.0,
         't_dim': 100001
     },
     'system': {
@@ -41,18 +40,16 @@ params = {
     },
     'plotter': {
         'type': 'contourf',
-        'palette': 'RdBu_r',
-        'bins': 11,
-        'title': '$\\delta = 0.01, \\eta = 0.75$',
+        'title': '$\\delta = 0.005, \\eta = 0.75$',
         'x_label': '$\\langle p_{L} \\rangle$',
-        'x_bound': 'both',
         'x_ticks': [-5, 0, 5],
         'y_label': '$\\langle p_{R} \\rangle$',
-        'y_bound': 'both',
         'y_ticks': [-5, 0, 5],
-        'cbar_ticks': [0, 1e-3],
-        'label_font_size': 33,
-        'tick_font_size': 27
+        'cbar_title': '$\\times 10^{-4}$',
+        'cbar_ticks': [0.0, 0.0004, 0.0008],
+        'cbar_tick_labels': [0.0, 0.4, 0.8],
+        'width': 5.5,
+        'height': 5.0
     }
 }
 
@@ -60,10 +57,10 @@ params = {
 init_log()
 
 # initialize system
-system = Uni00(params['system'])
+system = Uni00(params=params['system'])
 
-# get the dynamics
-M, _ = system.get_measure_dynamics(params['solver'], system.ode_func, system.get_ivc)
+# get correlation elements
+M, _ = system.get_measure_dynamics(solver_params=params['solver'])
 V_mech = [np.reshape(m, (4, 4)) for m in M[-1:]]
 
 # initialize variables
@@ -73,7 +70,7 @@ p_0s = np.around(np.linspace(-5, 5, 101), 1)
 p_1s = np.around(np.linspace(-5, 5, 101), 1)
 
 for V in V_mech:
-    # wigner
+    # wigner function
     W = list()
     for p_0 in p_0s:
         temp = list()
@@ -84,10 +81,9 @@ for V in V_mech:
         W.append(temp)
 
     # plotter
-    axes = {
+    plotter = MPLPlotter(axes={
         'X': p_0s.tolist(),
         'Y': p_1s.tolist()
-    }
-    plotter = MPLPlotter(axes, params['plotter'])
+    }, params=params['plotter'])
     plotter.update(vs=W)
-    plotter.show(True, 6.0, 5.0)
+    plotter.show(True)