+++ systems/Uni00.py
@@ -1,11 +1,11 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
  
-"""Class to simulate a simple unidirectionally-coupled configuration of QOM systems."""
+"""Class to simulate a simple unidirectionally-coupled configuration of two QOM systems."""
 
 __authors__ = ['Sampreet Kalita']
 __created__ = '2020-01-04'
-__updated__ = '2021-07-07'
+__updated__ = '2021-10-20'
 
 # dependencies
 import numpy as np
@@ -14,24 +14,26 @@ import numpy as np
 from qom.systems import DODMSystem
 
 class Uni00(DODMSystem):
-    """Class to simulate a simple unidirectionally-coupled configuration of QOM systems.
+    """Class to simulate a simple unidirectionally-coupled configuration of two QOM systems.
 
     Parameters
     ----------
     params : dict
         Parameters for the system.
+    cb_update : callable, optional
+        Callback function to update status and progress, formatted as ``cb_update(status, progress, reset)``, where ``status`` is a string, ``progress`` is an integer and ``reset`` is a boolean.
     """
 
-    def __init__(self, params):
+    def __init__(self, params, cb_update=None):
         """Class constructor for Uni00."""
         
         # initialize super class
-        super().__init__(params)
+        super().__init__(params=params, cb_update=cb_update)
 
         # set attributes
         self.code = 'uni_00'
-        self.name = 'Unidirectionally-coupled Configuration'
-        # set parameters
+        self.name = 'Two Unidirectionally-coupled QOM Systems'
+        # default parameters
         self.params = {
             'A_l': params.get('A_l', 52.0),
             'Delta_0': params.get('Delta_0', 1.0),
@@ -43,10 +45,11 @@ class Uni00(DODMSystem):
             'n_ths': params.get('n_ths', [0, 0]),
             'omega_m': params.get('omega_m', 1.0),
         }
-        # drift matrix
+        
+        # matrices
         self.A = None
 
-    def get_A(self, modes, params, t):
+    def get_A(self, modes, params, t=None):
         """Function to obtain the drift matrix.
 
         Parameters
@@ -55,8 +58,8 @@ class Uni00(DODMSystem):
             Values of the modes.
         params : list
             Constant parameters.
-        t : float
-            Time at which the rates are calculated.
+        t : float, optional
+            Time at which the drift matrix is calculated.
         
         Returns
         -------
@@ -75,11 +78,9 @@ class Uni00(DODMSystem):
         betas   = [modes[1], modes[3]]
         temp    = np.sqrt(eta * kappas[0] * kappas[1])
 
-        # initialize lists
+        # effective values
         Deltas  = list()
         gs      = list()
-
-        # effective detunings
         for i in range(2):
             Deltas.append(Delta_0s[i] + 2 * g_0s[i] * np.real(betas[i]))
             gs.append(g_0s[i] * alphas[i])
@@ -88,21 +89,23 @@ class Uni00(DODMSystem):
         if self.A is None or np.shape(self.A) != (8, 8):
             self.A = np.zeros([8, 8], dtype=np.float_)
         for i in range(2):
+            # X quadratures
             self.A[4*i + 0][4*i + 0] = - kappas[i]
             self.A[4*i + 0][4*i + 1] = - Deltas[i]
             self.A[4*i + 0][4*i + 2] = - 2 * np.imag(gs[i])
-
+            # Y quadratures
             self.A[4*i + 1][4*i + 0] = Deltas[i]
             self.A[4*i + 1][4*i + 1] = - kappas[i]
             self.A[4*i + 1][4*i + 2] = 2 * np.real(gs[i])
-
+            # Q quadratures
             self.A[4*i + 2][4*i + 2] = - gammas[i]
             self.A[4*i + 2][4*i + 3] = omega_ms[i]
-
+            # P quadratures
             self.A[4*i + 3][4*i + 0] = 2 * np.real(gs[i])
             self.A[4*i + 3][4*i + 1] = 2 * np.imag(gs[i])
             self.A[4*i + 3][4*i + 2] = - omega_ms[i]
             self.A[4*i + 3][4*i + 3] = - gammas[i]
+        # unidirectional terms
         self.A[4][0] = - 2 * temp
         self.A[5][1] = - 2 * temp
 
@@ -190,7 +193,7 @@ class Uni00(DODMSystem):
 
         return iv, c
 
-    def get_mode_rates(self, modes, params, t):
+    def get_mode_rates(self, modes, params, t=None):
         """Function to obtain the rates of the optical and mechanical modes.
 
         Parameters
@@ -199,7 +202,7 @@ class Uni00(DODMSystem):
             Values of the modes.
         params : list
             Constants parameters.
-        t : float
+        t : float, optional
             Time at which the rates are calculated.
         
         Returns
@@ -219,18 +222,18 @@ class Uni00(DODMSystem):
         alphas  = [modes[0], modes[2]]
         betas   = [modes[1], modes[3]]
 
-        # initialize lists
+        # effective values
         Deltas      = list()
         gs          = list()
-        dalpha_dts  = list()
-        dbeta_dts   = list()
-
-        # effective detunings
         for i in range(2):
             Deltas.append(Delta_0s[i] + 2 * g_0s[i] * np.real(betas[i]))
             gs.append(g_0s[i] * alphas[i])
 
-        # calculate rates
+        # initialize lists
+        dalpha_dts  = list()
+        dbeta_dts   = list()
+
+        # calculate mode rates
         for i in range(2):
             dalpha_dts.append((- kappas[i] + 1j * Deltas[i]) * alphas[i])
             dbeta_dts.append(1j * gs[i] * np.conjugate(alphas[i]) + (- gammas[i] - 1j * omega_ms[i]) * betas[i])